Utils: improve request rate measuring logic
As we are measuring the request rate in some deta time intervals,
we should not consider the corrent count to report the rate;
instead, this counter is only valid when the delta time is closed.
so we have to keep the previous counter and swap it with the
incrementing one, when the time is reached.